Dropping callback-based API; only use Promises
Background: NodeJS currently warns (and will terminate the process
in future releases) when a promise becomes unrejected, but this re-
jection is not handled. This makes it impossible to provide a hybrid
API that lets you mix-n-match between Promises and callback at will.

This forces us to use Promises, but its 2017 afterall any polyfills for
Promises are widely available, and every modern browser ships with
native support for that.